requirements.txt
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-dotenv>=1.0.0
pytest>=7.4.0

